+++ src/cdi_analytic/Analytic_Gray_Opacity.cc
@@ -256,7 +256,7 @@ Analytic_Gray_Opacity::sf_char Analytic_Gray_Opacity::pack() const
     packer.set_buffer(size, &packed[0]);
 
     // pack the anal_model size
-    packer << anal_model.size();
+    packer << static_cast<int>(anal_model.size());
 
     // now pack the anal model
     for (int i = 0; i < anal_model.size(); i++)
